+++ test/domain_test.cpp
@@ -15,7 +15,7 @@ TEST(DomainTest, Constants) {
 
 // NOLINTBEGIN(readability-function-cognitive-complexity)
 TEST(DomainTest, SubscriberIdManualCheck) {
-  for (const uint8_t num : {0, 65, 128, 255}) {
+  for (const uint8_t num : {(uint8_t)0, (uint8_t)65, (uint8_t)128, (uint8_t)255}) {
     ASSERT_THROW({ std::ignore = SubscriberId::create(num); }, std::invalid_argument);
     ASSERT_THROW({ std::ignore = SubscriberId::all_subscribers_mask(num); }, std::invalid_argument);
   }
@@ -24,7 +24,7 @@ TEST(DomainTest, SubscriberIdManualCheck) {
     ASSERT_EQ(0, sub_id.index());
     ASSERT_EQ(0b1, sub_id.mask());
 
-    const uint32_t mask = SubscriberId::all_subscribers_mask(1);
+    const uint64_t mask = SubscriberId::all_subscribers_mask(1);
     ASSERT_EQ(0b1, mask);
   }
   {
@@ -32,7 +32,7 @@ TEST(DomainTest, SubscriberIdManualCheck) {
     ASSERT_EQ(1, sub_id.index());
     ASSERT_EQ(0b10, sub_id.mask());
 
-    const uint32_t mask = SubscriberId::all_subscribers_mask(2);
+    const uint64_t mask = SubscriberId::all_subscribers_mask(2);
     ASSERT_EQ(0b11, mask);
   }
   {
@@ -40,7 +40,7 @@ TEST(DomainTest, SubscriberIdManualCheck) {
     ASSERT_EQ(2, sub_id.index());
     ASSERT_EQ(0b100, sub_id.mask());
 
-    const uint32_t mask = SubscriberId::all_subscribers_mask(3);
+    const uint64_t mask = SubscriberId::all_subscribers_mask(3);
     ASSERT_EQ(0b111, mask);
   }
   {
@@ -48,7 +48,7 @@ TEST(DomainTest, SubscriberIdManualCheck) {
     ASSERT_EQ(3, sub_id.index());
     ASSERT_EQ(0b1000, sub_id.mask());
 
-    const uint32_t mask = SubscriberId::all_subscribers_mask(4);
+    const uint64_t mask = SubscriberId::all_subscribers_mask(4);
     ASSERT_EQ(0b1111, mask);
   }
   {
@@ -56,7 +56,7 @@ TEST(DomainTest, SubscriberIdManualCheck) {
     ASSERT_EQ(30, sub_id.index());
     ASSERT_EQ(0b1000000000000000000000000000000, sub_id.mask());
 
-    const uint32_t mask = SubscriberId::all_subscribers_mask(31);
+    const uint64_t mask = SubscriberId::all_subscribers_mask(31);
     ASSERT_EQ(0b1111111111111111111111111111111, mask);
   }
   {
@@ -64,7 +64,7 @@ TEST(DomainTest, SubscriberIdManualCheck) {
     ASSERT_EQ(31, sub_id.index());
     ASSERT_EQ(0b10000000000000000000000000000000, sub_id.mask());
 
-    const uint32_t mask = SubscriberId::all_subscribers_mask(32);
+    const uint64_t mask = SubscriberId::all_subscribers_mask(32);
     ASSERT_EQ(0b11111111111111111111111111111111, mask);
   }
 
@@ -99,7 +99,7 @@ TEST(DomainTest, AllSubscribersMask) {
     if (num > ShmSequencer::MAX_SUBSCRIBER_NUM || num == 0) {
       ASSERT_THROW({ std::ignore = SubscriberId::all_subscribers_mask(num); }, std::invalid_argument);
     } else {
-      const uint32_t all_mask = SubscriberId::all_subscribers_mask(num);
+      const uint64_t all_mask = SubscriberId::all_subscribers_mask(num);
       ASSERT_EQ(pow(2, num) - 1, all_mask);
     }
   });
